Allow to cancel quitting with ^X.
When quitting is blocked by unsaved files, allows to abort the quit by a
starting a ^X key sequence, without first dismissing the "Quit?"
question. The primary expected use is for the user to do ^X-^S, ^X-^C
when realizing that their buffer wasn't saved.

(Typing a 'n' or hitting return or enter before the ^X-^S may seem like
a small step, but my own experience is that I forget and want to go
straight to ^X-^s.)