When wiping last line, don't add newline.
The fixes a weirdness where wiping the last line is not undone by a
single ^Z, due to the newline insertion being a separate event in the
undo action list (b.ual). Thus, to get back to the state prior to a wipe
of the last line, two undos were required.

With this modification, the behavior when wiping the last line is
slightly modified. Specifically, the wipe does not include a newline at
the end (after all, there was no newline to wipe in the first place). As
a consequence, when yanking the wiped line, there we will also be no
newline at the end of the insert.

For precedence, this is consistent with the emacs ^K on the last line
(even if wipe behavior is not in general identical to ^K).